// src/app/services/models/ProjectToReturn.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Image } from '../models/image';
export interface ProjectToReturn {
  description?: string;
  domaine?: 'Web' | 'CyberSecurity' | 'Ai_Models';
  image?: Image;
  isOpen?: boolean;
  isVisible?: boolean;
  name?: string;
  professorFullName?: string;
  studentFullName?: string;
  createdAt?:string;
}